backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
SQLAlchemy[asyncio]==2.0.36
asyncpg==0.29.0
alembic==1.13.2
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==3.2.2
python-multipart==0.0.9
aiofiles==23.2.1
httpx==0.27.2
apscheduler==3.10.4
slowapi==0.1.9
loguru==0.7.2
email-validator==2.2.0
pytest==8.3.3
pytest-asyncio==0.24.0
